refactor: consolidate job applications in data store
